% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates_keeplast.R
\name{filter_duplicates_keeplast}
\alias{filter_duplicates_keeplast}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_duplicates_keeplast(df, group_vars = c("user_id", "Session"),
  time_var = "End Time")
}
\arguments{
\item{df}{class: numeric; original X}

\item{group_vars}{class: vector; factors to group data by}

\item{time_var}{class: vector; variable for time}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
filter_duplicates_keeplast(df, group_vars=c("user_id", "Session"), time_var="End Time")
}
\keyword{cognition}
\keyword{m2c2,}
