% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negate.R
\name{`\%nin\%`}
\alias{`\%nin\%`}
\alias{\%nin\%}
\title{RM2C2: Scoring, Summarizing}
\usage{
... \%nin\% NA
}
\arguments{
\item{`%in%`}{class: object to negate}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
\%nin\%(x)
}
\keyword{cognition}
\keyword{m2c2,}
